+++ src/python/pants/jvm/compile_test.py
@@ -46,10 +46,10 @@ from pants.jvm.compile import (
     ClasspathSourceMissing,
 )
 from pants.jvm.goals import lockfile
+from pants.jvm.resolve.common import CoursierResolveKey
 from pants.jvm.resolve.coursier_fetch import CoursierFetchRequest
 from pants.jvm.resolve.coursier_fetch import rules as coursier_fetch_rules
 from pants.jvm.resolve.coursier_setup import rules as coursier_setup_rules
-from pants.jvm.resolve.key import CoursierResolveKey
 from pants.jvm.target_types import JvmArtifactTarget
 from pants.jvm.testutil import (
     RenderedClasspath,
