+++ src/python/pants/jvm/resolve/coursier_fetch_test.py
@@ -12,10 +12,9 @@ from pants.backend.java.target_types import DeployJarTarget, JavaSourcesGenerato
 from pants.backend.java.target_types import rules as target_types_rules
 from pants.core.util_rules import config_files, source_files
 from pants.engine.addresses import Address, Addresses
-from pants.jvm.resolve.common import Coordinate
+from pants.jvm.resolve.common import Coordinate, CoursierResolveKey
 from pants.jvm.resolve.coursier_fetch import NoCompatibleResolve
 from pants.jvm.resolve.coursier_fetch import rules as coursier_fetch_rules
-from pants.jvm.resolve.key import CoursierResolveKey
 from pants.jvm.target_types import JvmArtifactTarget
 from pants.jvm.testutil import maybe_skip_jdk_test
 from pants.jvm.util_rules import rules as util_rules
